+++ src/main/java/org/lushplugins/lushrewards/module/dailyrewards/DailyRewardsModule.java
@@ -7,7 +7,6 @@ import org.lushplugins.lushrewards.gui.GuiDisplayer;
 import org.lushplugins.lushrewards.gui.GuiFormat;
 import org.lushplugins.lushrewards.module.RewardModule;
 import org.lushplugins.lushrewards.module.UserDataModule;
-import org.lushplugins.lushrewards.rewards.collections.DailyRewardCollection;
 import org.lushplugins.lushrewards.rewards.collections.RewardDay;
 import org.lushplugins.lushrewards.utils.Debugger;
 import org.lushplugins.lushlib.gui.inventory.Gui;
