+++ src/main/java/org/lushplugins/lushrewards/module/dailyrewards/DailyRewardsPlaceholder.java
@@ -1,7 +1,6 @@
 package org.lushplugins.lushrewards.module.dailyrewards;
 
 import org.lushplugins.lushrewards.LushRewards;
-import org.lushplugins.lushrewards.rewards.collections.DailyRewardCollection;
 import org.lushplugins.lushrewards.rewards.collections.RewardDay;
 import org.lushplugins.lushrewards.utils.LocalPlaceholders;
 import org.lushplugins.lushlib.module.Module;
